Addressed Prateek's comments and moved the tests as well
